added check for package

+++ inc/get_package.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 function get_package {
   package='puppet5-release-bionic.deb'
-  sudo wget -O /tmp/${package} https://apt.puppetlabs.com/${package}
-  sudo dpkg -i /tmp/${package}
+  # need a check if installed
+  if ! dpkg -l | grep puppet
+  then
+    # install package
+    sudo wget -O /tmp/${package} https://apt.puppetlabs.com/${package}
+    sudo dpkg -i /tmp/${package}
+  fi
 }
